refactor: move validateApiKey to service
